refactor: move token to utils
