+++ packages/nns/src/index.ts
@@ -2,7 +2,6 @@ export type { RewardEvent } from "../candid/governance";
 export type { DeployedSns } from "../candid/sns_wasm";
 export { AccountIdentifier, SubAccount } from "./account_identifier";
 export * from "./enums/governance.enums";
-export * from "./enums/token.enums";
 export * from "./errors/governance.errors";
 export * from "./errors/ledger.errors";
 export { GenesisTokenCanister } from "./genesis_token.canister";
@@ -10,7 +9,6 @@ export { GovernanceCanister } from "./governance.canister";
 export { ICP } from "./icp";
 export { LedgerCanister } from "./ledger.canister";
 export { SnsWasmCanister } from "./sns_wasm.canister";
-export * from "./token";
 export * from "./types/common";
 export * from "./types/governance.options";
 export * from "./types/governance_converters";
