+++ packages/utils/src/constants/constants.ts
@@ -1,0 +1,1 @@
+export const E8S_PER_TOKEN = BigInt(100000000);
